Move MyAdapter
